fix for Uncaught ReferenceError: p is not definer when expanding groups
In reference to issue #86 

Now that I have had a chance to test this out, I see that the issue has been resolved but now there is a new bug which doesn't allow an expanded group to close and reopen without throwing an error in the console: Uncaught ReferenceError: p is not defined. This push request fixed that error.